+++ src/main/java/com/project/phonecaseshop/repository/ModelRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 @Repository
 public interface ModelRepository extends JpaRepository<Model, Long> {
-    List<Model> findByProductId_ProductId(Long productId);
+    List<Model> findByProductId_ProductId(int productId);
 
 }
