+++ src/main/java/com/project/phonecaseshop/repository/RefreshTokenRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 
-    Optional<RefreshToken> findByMemberId(Long memberId);
+    Optional<RefreshToken> findByMemberId(int memberId);
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
 }
