+++ src/main/java/com/project/phonecaseshop/security/filter/JwtAuthenticationFilter.java
@@ -70,7 +70,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         Optional<RefreshToken> refreshTokenCheck = refreshTokenRepository.findByRefreshToken(refreshToken);
 
         if (refreshTokenCheck.isPresent()) {
-            Optional<Member> member = memberRepository.findById(refreshTokenCheck.get().getMemberId());
+            Optional<Member> member = memberRepository.findById((long) refreshTokenCheck.get().getMemberId());
             if (member.isPresent()) {
                 String reIssuedRefreshToken = reIssuedRefreshToken(refreshTokenCheck.get().getMemberId());
                 String reIssuedAccessToken = tokenUtil.createAccessToken(member.get().getMemberEmail());
@@ -79,7 +79,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         }
     }
 
-    public String reIssuedRefreshToken(Long memberId) {
+    public String reIssuedRefreshToken(int memberId) {
         String reIssuedRefreshToken = tokenUtil.createRefreshToken();
         tokenUtil.insertRefreshToken(memberId, reIssuedRefreshToken);
         return reIssuedRefreshToken;
